[Fix #292] 소켓 분리, nav 닉네임 나오게 수정
